refactor: move Legacy methods to Mesh_Legacy.cpp, add DNDS_TEST_NP_LIST env variable
Move GeneralCell2NodeToNode2Cell, RecoverNode2CellAndNode2BndLegacy, and
RecoverCell2CellAndBnd2CellLegacy to src/Geom/Mesh_Legacy.cpp (450 lines
extracted from Mesh.cpp). InterpolateFaceLegacy stays in Mesh.cpp since
it depends on shared anonymous-namespace helpers.

Add DNDS_TEST_NP_LIST environment variable to control which MPI rank
counts are tested. Read at CMake configure time, default '1;2;4;8'.
Replaces hardcoded np=1,2,4,8 blocks in all MPI test registration
functions (dnds, geom, cfv, euler) with a single foreach loop.

Usage: DNDS_TEST_NP_LIST='1;2' cmake --preset release-test
       # registers only np1 and np2 tests

18/18 geom tests pass at np=1,2,4,8 with default NP list.
